refactor(ui): extract ChatView turn rendering
