+++ packages/ui/src/components.tsx
@@ -8,7 +8,8 @@ export { describeLoadToolResult, formatRedactedJson, formatToolIntent, loadToolD
 export { formatBytes, OverlayHost, ToolActivity, ToolErrorDetails } from './tool-activity.js';
 export { PermissionDialog } from './permission-dialog.js';
 export { ChatView } from './chat-view.js';
-export type { ChatHeaderAlert, TurnFooterActionMeta, TurnLineageBadge } from './chat-view.js';
+export type { ChatHeaderAlert } from './chat-view.js';
+export type { TurnFooterActionMeta, TurnLineageBadge } from './chat-turn.js';
 export { AutomationsPage, DailyReviewPage, SkillsPage } from './module-pages.js';
 export { Composer } from './composer.js';
 export type { ComposerHandle } from './composer.js';
